+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: akroll <akroll@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/22 10:58:33 by akroll            #+#    #+#             */
-/*   Updated: 2022/03/22 14:25:54 by akroll           ###   ########.fr       */
+/*   Updated: 2022/03/24 12:24:35 by akroll           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 
-int ft_strlen(const char *s)
+size_t ft_strlen(const char *s)
 {
-    int len;
+    size_t len;
 
     len = 0;
     while (s[len] != '\0')
